Long-Term Authentication
This is the same methodology as the StreamElements OAuth2 system.
This permanently adds a file on my website as "required" to function.
The refresh happens no more than once per hour.
The refresh is triggered by the ~5 minute PING from Twitch's IRC server.
Expires is now essentially obsolete.
The active OAuth token and the current Refresh token are kept in Local Storage.
The Refresh token in Local Storage is erased on a 401 from the refresh.
Use of the old cfg.login.oauth value is still possible.