Add Tagged Channel Trigger
If you add an "@channel" tag to your stream title, "@"-ing that channel at the beginning of a message becomes a kappagen trigger.
The kappagen will not use emotes from the message, but instead either your channel's emotes or the target channel's emotes.
The access is controlled by an access value, but falls back to the general kappagen command's access value if less than 0.
Reorganized other message-based events a little to prevent possible bugs.