Allow Stretching Emotes to 256x256 by Default
This should allow bigger emote size ratios for mobile-friendly or comedically large emote sizes.
When this variable was set, The Cube kappagen hadn't been created, and I wanted to prevent emotes from being larger than the original image source. Since that went out the window with The Cube, we might as well allow a better range.