Ignore CORS and Unknown YT Chat Errors if Active
However, the actual unknown error still needs to be reported.

Note: "Active" is defined by the presence of pagination, which is removed in the retry, preventing a second retry. However, if that initial request succeeds and the next, paginated request fails in a loop for some reason, this could still result in a potential infinite loop of undetected failures - but if the initial successes return chat data, I suppose the point is moot.